Added pLow, pHigh, pClose & rsi

+++ AutoTrade/src/net/icegalaxy/Rules.java
@@ -9,7 +9,7 @@ public abstract class Rules implements Runnable
 	protected boolean hasContract;
 	protected double tempCutLoss;
 	protected double tempStopEarn;
-	protected float refPt;
+	protected double refPt;
 	protected float buyingPoint;
 	private boolean globalRunRule;
 	protected String className;
